Fix bugs after refactoring
