+++ src/components/task.js
@@ -1,6 +1,6 @@
 import AbstractComponent from './abstract-component';
 import {MonthNames} from '../const';
-import {formatTime} from '../utils';
+import {formatTime} from '../utils/common';
 
 const createHashtagsMarkup = (hashtags) => {
   return hashtags
